added SourceMessagePreview
