+++ lib/typescript/render/index.ts
@@ -1,0 +1,4 @@
+export * from './props';
+export * from './components/Avatar';
+export * from './SourceMessage';
+export * from './SourceMessagePreview';
